First stage in refactoring Bullet: moved Bullet Collision and Dynamics and LinearMath into src folder, and all files in Collision Detection and Dynamics have bt prefix. Made all buildsystems to work again (jam, msvc, cmake)

+++ Demos/BspDemo/BspConverter.cpp
@@ -16,7 +16,7 @@ subject to the following restrictions:
 #include "BspConverter.h"
 #include "BspLoader.h"
 #include "CcdPhysicsEnvironment.h"
-#include "SimdVector3.h"
+#include "LinearMath/SimdVector3.h"
 
 void BspConverter::convertBsp(BspLoader& bspLoader,float scaling)
 {
@@ -86,7 +86,7 @@ void BspConverter::convertBsp(BspLoader& bspLoader,float scaling)
 							std::vector<SimdVector3>	vertices;
 							
 							getVerticesFromPlaneEquations(planeEquations,vertices);
-							printf("getVerticesFromPlaneEquations returned %i\n",vertices.size());
+							printf("getVerticesFromPlaneEquations returned %i\n",(int)vertices.size());
 
 							bool isEntity = false;
 							SimdVector3 entityTarget(0.f,0.f,0.f);
